+++ packages/myst-cli/src/process/myst.ts
@@ -3,6 +3,7 @@ import { mystParse } from 'myst-parser';
 import { cardDirective } from 'myst-ext-card';
 import { gridDirective } from 'myst-ext-grid';
 import { proofDirective } from 'myst-ext-proof';
+import { exerciseDirectives } from 'myst-ext-exercise';
 import { reactiveDirective, reactiveRole } from 'myst-ext-reactive';
 import { tabDirectives } from 'myst-ext-tabs';
 import { VFile } from 'vfile';
@@ -14,7 +15,14 @@ export function parseMyst(session: ISession, content: string, file: string): Roo
   vfile.path = file;
   const parsed = mystParse(content, {
     markdownit: { linkify: true },
-    directives: [cardDirective, gridDirective, reactiveDirective, proofDirective, ...tabDirectives],
+    directives: [
+      cardDirective,
+      gridDirective,
+      reactiveDirective,
+      proofDirective,
+      ...exerciseDirectives,
+      ...tabDirectives,
+    ],
     roles: [reactiveRole],
     vfile,
   });