+++ packages/myst-templates/src/types.ts
@@ -13,8 +13,12 @@ export type ValueAndIndex = {
   letter: string;
 };
 
-export type RendererAuthor = Omit<Author, 'affiliations' | 'corresponding' | 'orcid'> & {
+export type RendererAuthor = Omit<
+  Author,
+  'affiliations' | 'collaborations' | 'corresponding' | 'orcid'
+> & {
   affiliations?: ValueAndIndex[];
+  collaborations?: ValueAndIndex[];
   corresponding?: ValueAndIndex;
   orcid?: string;
   index: number;
@@ -31,9 +35,10 @@ export type RendererDoc = Omit<PageFrontmatter, 'date' | 'authors'> & {
   };
   authors: RendererAuthor[];
   affiliations: ValueAndIndex[];
+  collaborations: ValueAndIndex[];
 };
 
-export const RENDERER_DOC_KEYS = ['affiliations'].concat(PAGE_FRONTMATTER_KEYS);
+export const RENDERER_DOC_KEYS = ['affiliations', 'collaborations'].concat(PAGE_FRONTMATTER_KEYS);
 
 export type TemplatePartDefinition = {
   id: string;