+++ packages/myst-roles/CHANGELOG.md
@@ -1,5 +1,7 @@
 # myst-roles
 
+## 0.0.31
+
 ## 0.0.30
 
 ### Patch Changes